Some refactoring made.
